refactor(server): split util and finance to separate files
